edge-node/requirements.txt
Flask
opencv-python
Pillow
python-socketio
websocket-client
python-dotenv
numpy
requests

